[Feature:InstructorUI] New Toggle Column Options (#9929)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
The columns in the Manage Students page is missing columns for
additional data we have on students. The only way to view this
information is to press the "Edit Student" button for each individual
student.


![image](https://github.com/Submitty/Submitty/assets/92200885/4879a7c0-93c7-436d-9caf-9da09a39f47e)

### What is the new behavior?
There are five new column options in the Manage Students page when the
Toggle Columns button is pressed. When checked off there will be a new
column for the additional data showing the information for all students
instead of one. The new options are turned off by default.


![image](https://github.com/Submitty/Submitty/assets/92200885/6f7884dc-ce6d-461a-ab5e-90a6050a4132)
(Image is cut off, you can scroll left/right to see all the columns on
the site.)

![image](https://github.com/Submitty/Submitty/assets/92200885/614cb941-66ba-42d4-9a46-7f00c57befd3)


### Other information?
First task in #9848