[Refactor:Forum] Replace iterated database queries (#9988)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
Closes #9981 Currently, loading the forum pages require executing loops
and querying the database from within the loops. This resulted in
hundreds of repetitive queries, slowing the page down.

### What is the new behavior?
Now the forum makes a few queries outside of loops, then iterates over
the results. This will make the forum much faster.

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->
The main functions that render the forum are displayThreadList (for the
main forum page, lists all the threads) and generatePostList (displays
the posts in the current thread). Queries inside the large for loop in
displayThreadList have been moved outside, and the main forum page now
requires 15 queries to load. generatePostList calls createPost in a
loop, so I queried the data needed in createPost outside the loop in
generatePostList. The thread pages are now down to 31-32 queries.

Moving these queries also caused some old query functions to never be
called, so I deleted those. I also made the change to showThreads so I
could delete getResolveState (the data provided is already queried by
getThread).

---------

Co-authored-by: Barb Cutler <bmcutler@users.noreply.github.com>