[Feature:CourseMaterials] Button to clear "NEW" badge  (#10032)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
New materials posted in the course materials page show up with a "NEW"
badge next to them. However, there is no way to mark all the course
materials as seen at once. The only way to mark them as viewed is to
download them which can be an inconvenience if there are a lot of small
files that don't need to be viewed.

### What is the new behavior?
There is now a button that marks all of the course materials as seen,
which gets rid of the "NEW" badge next to them.

![image](https://github.com/Submitty/Submitty/assets/92200885/b49835dc-749b-4085-8eeb-2a948f7e52d4)


![image](https://github.com/Submitty/Submitty/assets/92200885/470614ed-3375-4851-bf26-3c6729df49f7)


### Other information?
Closes issue #9941

---------

Co-authored-by: Chuanyi Pan <cy2631995926@gmail.com>
Co-authored-by: Chris Reed <55092742+cjreed121@users.noreply.github.com>
Co-authored-by: Chuanyi Pan <92846547+VVVITH@users.noreply.github.com>