[Bugfix:Forum] Highlight new posts in Forum (#10324)
### What is the current behavior?
When a user got a new post while he is in the forum page, user's browser
will render it as a read message. But, it should be rendered as an
unread message.

### What is the new behavior?
Now, if a user got a post while user is in the forum page, browser will
render it as an unread message.

### Other information?
(Changed according to the request from @DarthNyan.)
Changes : Added a code line to add 'new_post' class when displaying a
new post.
I tested this on my VM.
It works on dark mode as well (a white color transparent overlay). 

However, I was unable to find the logical error why the system treat new
messages like this. Let's try to find a more sustainable method soon.
Working on it :)

closes #10319 

Thank you.