[Feature:Calendar] Changed Calendar Cells Based on View (#10281)
Changed the calendar cells to be larger based on the viewing mode(month,
two-week, week)

![Screenshot from 2024-03-25
15-29-22](https://github.com/Submitty/Submitty/assets/122404476/11e7597f-2ecf-4b10-9777-b66791b91591)

![Screenshot from 2024-03-25
15-29-09](https://github.com/Submitty/Submitty/assets/122404476/ba3ea95b-28e8-4981-8e32-1bf2c68e5240)

The current behavior is that if you change viewing modes from 'Month' to
'Two Week' or 'Week', it simply hides the extra weeks. The purpose of
the other view modes is to make cells larger to fit more information and
be less crammed for users who have many items on their calendar.

The new behavior is that the cells will expand based on the viewing mode
as well as the number of items in each cell.

This is not a breaking change and was tested on instructor and student
accounts.