[Bugfix:Forum] Disable Forum Reply Button (#10129)
### Please check if the PR fulfills these requirements:

* [x] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [x] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
Fixes [#10126](https://github.com/Submitty/Submitty/issues/10126)
When a user is currently on a forum thread, any reply button is always
enabled, regardless of the empty text area above. This can be misleading
to users and it would be better if reply buttons were always disabled on
empty text area input.


### What is the new behavior?
The reply buttons are now always set to disabled whenever the text area
is empty. This is checked on page load, thread change, and any change in
text area input. This also applies to replying to any thread block or
even editing your post on the forum.

### Other information?
This is a small change. Should be working as intended as it was tested
on student and instructor accounts. Changes were tested for posts
created by current users and posts created by others.


![image](https://github.com/Submitty/Submitty/assets/94331560/7a939aeb-a0b0-44a0-b2f0-45660c46472d)


![image](https://github.com/Submitty/Submitty/assets/94331560/f5b763d5-eb34-4333-ac96-b31cdd80aa29)

---------

Co-authored-by: Josh <whb20050728@gmail.com>