[Bugfix:Forum] Fix forum socket update (#10388)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
If you are on the forum page and a new post is added through the
websocket, an error will come up that it cannot parse new post. It will
also frog robot on the backend.

### What is the new behavior?
New posts will dynamically load in without an issue. This was due to
missing arguments in the createPost function when called by
renderTemplate.

### Other information?
Tested locally and new forum posts load fine