[Refactor:Calendar] Luxon Date Objects (#9989)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [ ] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
Fixes #9731
We use the builtin JS date library in a number of places. While this
mostly works, we keep getting bit by browser implementation specific
stuff and pitfalls because it's not a very developer friendly interface.
This is especially a bit hairy as we then do bits of math with these
dates, which is equally error prone. Many of the strings passed in do
not follow the ISO8601 format and rely on non-standard browser behavior
to parse the date.
### What is the new behavior?
Instead of using the builtin JS date library, we will be using Luxon. I
have only made this change in latedays.js so the team can review and
approve of integrating Luxon for the rest of Submitty.
### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->

---------

Co-authored-by: Youssef-Chip <131885371+Youssef-Chip@users.noreply.github.com>
Co-authored-by: Barb Cutler <bmcutler@users.noreply.github.com>
Co-authored-by: DarthNyan <tatewhiteberg@gmail.com>
Co-authored-by: Tate Whiteberg <89590361+DarthNyan@users.noreply.github.com>