[Bugfix:Submission] Errors in VCS submission (#10316)
### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
Fixes #10270

The value of `num_submission_boxes` is set to the number of parts in the
submission, but although a VCS submission has parts, it does not have
any submission boxes. This leads to errors and undefined behavior as the
page tries to access submission boxes that don't exist, making parts of
the page unusable for instructors as described in the linked issue.

### What is the new behavior?

The value of this property is set to 0 if the submission type is VCS,
allowing for the page to remain functional.

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->
Tested for both VCS and non-VCS submissions, and ensured that other
related functionality such as deleting and restoring files in submission
box still work for non-VCS submissions.