[Bugfix:Submission] Fix custom repo name bug (#10431)
### What is the current behavior?
If a gradeable is made with a custom repo name, a student won't be able
to submit due to a new check recently added. The generate repos script
sets the permission to daemoncgi which is not correct now.
Subdirectories were pulled from the database which would not work in the
case of a custom repo name.

### What is the new behavior?
The check has been modified to pick up custom repo names too so students
can submit or create the repo if needed correctly. The generate repos
script now sets permissions to daemonphpcgi. Subdirectories are now
another argument that can be specified of the generate repos script.

### Other information?
Tested locally and was able to create repos and submit after creation.

---------

Co-authored-by: Barb Cutler <bmcutler@users.noreply.github.com>