[Feature:System] Ansible Course Creation Scripts (#10488)
### Please check if the PR fulfills these requirements:

* [ ] Tests for the changes have been added/updated (if possible)
* [x] Documentation has been updated/added if relevant
* [ ] Screenshots are attached to Github PR if visual/UI changes were
made

### What is the current behavior?
<!-- List issue if it fixes/closes/implements one using the "Fixes
#<number>" or "Closes #<number>" syntax -->
Currently, there is only one method for creating a course and creating a
term, both following extensive manual process through command-line
prompts.

### What is the new behavior?

With the Ansible Term and Course Creation Scripts, it will allow for
using the Ansible automation tool to create a term and create a course,
all in one playbook. Editing that playbook will allow for more
customizable creation of the term and courses. This is a add on to the
Ansible Install Script that we have, all in the .setup/ansible
directory.

### Other information?
<!-- Is this a breaking change? -->
<!-- How did you test -->

---------

Co-authored-by: Preston Carman <prestonc@apache.org>
Co-authored-by: Cameron Peterson <46759635+IDzyre@users.noreply.github.com>