[UI/UX:InstructorUI] Add margin to error text (#10349)
### What is the current behavior?
There is no left-padding between h4 element and error-message
(span.red-message). That makes the UI unclean.

### What is the new behavior?

![image](https://github.com/Submitty/Submitty/assets/68075099/f40a08ac-e901-4ddc-8dae-4573dd62fccc)
There is a 5px of left-padding and the UI looks better.
closes #10344 

### Other information?
Changed electronic.css -> added a new style to elements with red-message
class where parent element is [#autograding_results .box-block] which is
where the auto grading error is displayed. So, other files will not be
affected that use "red-message" class.

---------

Co-authored-by: Barb Cutler <bmcutler@users.noreply.github.com>
Co-authored-by: Michael Papadopoulos <micpap25@gmail.com>
Co-authored-by: Zheyu Deng <122962727+ZheyuDeng@users.noreply.github.com>