[Bugfix:InstructorUI] Unload course after checking for rejoin (#10568)
### What is the current behavior?
Fixes #10565. When a student loses access to a course, their Submitty
front page shows the sidebar for the course they no longer have access
to. This is because Submitty checks if students can rejoin a course
they're no longer part of, this rejoin check loads the course config,
and the course config is never unloaded, so the sidebar is rendered.

### What is the new behavior?
The fix is to simply unload the course config after checking if the
student is permitted to rejoin.

### Other information?
Tested by removing and re-adding students to the Sample course in the
VM, and checking their front pages.