[Feature:Developer] Add CI flag to setup scripts (#10640)
### What is the current behavior?
Vagrant up jobs are hanging at the setup of NTP because of a 'soliciting
pool' infinite loop, which seemingly only happens in CI.
Currently we don't really have a way to tell if the Vagrant Up job is
happening in CI from within the setup scripts.

### What is the new behavior?
The `-c` optional parameter has been added to the setup_vagrant.sh file
. If you have the `CI=1` environment variable set, and you run vagrant
up, it will assume that you are running it in CI, and will pass the `-c`
flag to setup_vagrant.sh. This creates a file in the root directory of
Submitty, with the contents `This file is used to check if Submitty is
being run in the Github Actions CI.`. Later in the setup scripts, you
can test for the existence of this file, and therefor detect if the
script is being run in CI, and skip some optional, or breaking steps to
the setup process.

### Other information?
This can be tested locally by setting the environment variable CI=1, and
running `vagrant up`. You should notice soon after the VM is setup, and
all folders have been synced a message saying `Submitty is being run in
the Github Actions CI.`

https://github.com/IDzyre/Submitty/actions/runs/9707159611/job/26791907078
-- old way
https://github.com/IDzyre/Submitty/actions/runs/9718510404 show this
works